Dotfiles update
