+++ .cargo/config.toml
@@ -1,0 +1,13 @@
+[registries.crates.io]
+protocol = "sparse"
+
+[target.x86_64-unknown-linux-gnu]
+linker = "clang"
+rustflags = [
+    "-Clink-arg=-fuse-ld=lld",
+    "-Zshare-generics=y"
+]
+
+[target.x86_64-pc-windows-msvc]
+linker = "rust-lld.exe"
+rustflags = ["-Zshare=generics=n"]
