+++ .gitignore
@@ -1,1 +1,6 @@
+# mdbook
+/book
+
+# Added by cargo
 /target
+Cargo.lock
