+++ die4u-lib/src/core/animation/resources.rs
@@ -1,0 +1,6 @@
+use bevy::prelude::*;
+use bevy::time::Timer;
+
+/// Animation timer for entities that does not have their own timers
+#[derive(Resource)]
+pub struct GlobalSpriteAnimationTimer(pub Timer);
