+++ die4u-lib/src/core/assets/progress.rs
