+++ die4u-lib/src/core/setup/camera_setup.rs
