+++ die4u-lib/src/mobs/components/health.rs
