+++ die4u-lib/src/mobs/player/resources.rs
@@ -1,0 +1,7 @@
+use bevy::prelude::Resource;
+
+#[derive(Resource)]
+pub struct PlayerConfig {
+    pub speed: f32,
+    pub max_speed: f32,
+}
